Models renamed to proper laravel conventions
