+++ app/Models/Jobissue.php
