+++ app/Policies/JobPolicy.php
@@ -2,7 +2,7 @@
 
 namespace App\Policies;
 
-use App\Models\Jobs;
+use App\Models\Job;
 use App\Models\User;
 
 use Illuminate\Auth\Access\HandlesAuthorization;
@@ -12,11 +12,11 @@ class JobsPolicy
 {
     public function noActiveJobs()
     {
-        return Jobs::query()->currentUser()->activeJobs()->count()
+        return Job::query()->currentUser()->activeJobs()->count()
             ? Response::deny("Not allowed while a job is active") : Response::allow();
     }
 
-    public function view(?User $user, Jobs $job)
+    public function view(?User $user, Job $job)
     {
         return $user->id === $job->user_id;
     }
