+++ app/Policies/JobissuePolicy.php
@@ -2,8 +2,8 @@
 
 namespace App\Policies;
 
-use App\Models\Jobissues;
-use App\Models\Jobitems;
+use App\Models\Jobissue;
+use App\Models\Jobitem;
 use App\Models\User;
 use Illuminate\Auth\Access\Response;
 
@@ -20,7 +20,7 @@ class JobissuesPolicy
     /**
      * Determine whether the user can view the model.
      */
-    public function view(User $user, Jobissues $jobissues): bool
+    public function view(User $user, Jobissue $jobissues): bool
     {
         return $user->id == $jobissues->user_id;
     }
@@ -28,7 +28,7 @@ class JobissuesPolicy
     /**
      * Determine whether the user can create models.
      */
-    public function create(User $user, Jobissues $itemid): bool
+    public function create(User $user, Jobissue $itemid): bool
     {
         return $user->id == $itemid->job->user_id;
     }
@@ -36,7 +36,7 @@ class JobissuesPolicy
     /**
      * Determine whether the user can update the model.
      */
-    public function update(User $user, Jobissues $jobissues): bool
+    public function update(User $user, Jobissue $jobissues): bool
     {
         return false;
     }
@@ -44,7 +44,7 @@ class JobissuesPolicy
     /**
      * Determine whether the user can delete the model.
      */
-    public function delete(User $user, Jobissues $jobissues): bool
+    public function delete(User $user, Jobissue $jobissues): bool
     {
         return false;
     }
@@ -52,7 +52,7 @@ class JobissuesPolicy
     /**
      * Determine whether the user can restore the model.
      */
-    public function restore(User $user, Jobissues $jobissues): bool
+    public function restore(User $user, Jobissue $jobissues): bool
     {
         return false;
     }
@@ -60,7 +60,7 @@ class JobissuesPolicy
     /**
      * Determine whether the user can permanently delete the model.
      */
-    public function forceDelete(User $user, Jobissues $jobissues): bool
+    public function forceDelete(User $user, Jobissue $jobissues): bool
     {
         return false;
     }
